split test_basic_2
