+++ python/ray/tests/BUILD
@@ -26,6 +26,7 @@ py_test_module_list(
     "test_basic_3.py",
     "test_basic_4.py",
     "test_basic_5.py",
+    "test_basic_6.py",
     "test_cancel.py",
     "test_gcs_fault_tolerance.py",
     "test_metrics_agent.py",
@@ -325,6 +326,7 @@ py_test_module_list(
     "test_basic_3.py",
     "test_basic_4.py",
     "test_basic_5.py",
+    "test_basic_6.py",
     "test_asyncio.py",
     "test_multiprocessing.py",
     "test_list_actors.py",
